refactor(tui): split cover_art into pool, quantize, and derived_cache modules
cover_art.rs had grown to ~1,200 lines mixing four concerns. Split it
into a directory module: pool.rs (the background worker pool),
quantize.rs (half-block color extraction), derived_cache.rs (the
generic artifact lifecycle, with its tests alongside it), and mod.rs
(the CoverArtCache facade and TuiCoverArt client data, re-exporting
the public quantize items). Visibility is restricted to pub(super)
for the internals so the module boundary is enforced by the compiler.

No behavioral changes.
